FR #0006248

+++ branches/mgallery/webEdition/liveUpdate/sqldumps/tblthumbnails.sql
@@ -8,7 +8,7 @@ CREATE TABLE ###TBLPREFIX###tblthumbnails (
   Ratio tinyint(1) unsigned NOT NULL default '0',
   Maxsize tinyint(1) unsigned NOT NULL default '0',
   Interlace tinyint(1) unsigned NOT NULL default '1',
-  Fitinside smallint(5) unsigned NOT NULL default '0',
+  Fitinside tinyint(5) unsigned NOT NULL default '0',
   `Directory` varchar(1024) NOT NULL default '',
   Utilize tinyint(1) unsigned NOT NULL default '0',
   `Quality` tinyint unsigned NOT NULL DEFAULT  '8',
